[#10760] rename _dev._scm_version -> _dev.scm_version
